negative must not be empty (#450)
