+++ git-refspec/tests/fixtures/generated-archives/make_baseline.tar.xz
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:ace43ca904298dc76951863d41485ae0c21ee9b88abf7d604eb6c6e7efa18ab2
-size 9352
+oid sha256:6d2478d9f90c1e68924743a45cf801b4420924751c2caf7fa9a6c14291c388cd
+size 9356
